feat(example/utils): Move http_framework to dedicated file
